+++ themes/papirus/.gitignore
@@ -3,3 +3,5 @@ resume.json
 build/
 *.DS_Store
 public/
+src/templates.cjs
+src/partials.cjs
