+++ themes/papirus/package.json
@@ -7,16 +7,14 @@
     "type": "git",
     "url": "https://github.com/konalexiou/jsonresume-theme-papirus"
   },
-  "main": "build/index.js",
-  "exports": {
-    "next": "./src/index.js",
-    "default": "./build/index.js"
-  },
+  "main": "src/index.js",
   "scripts": {
     "lint": "eslint .",
-    "build": "webpack --mode production",
-    "prepack": "turbo build"
+    "build": "handlebars ./partials -p -e hbs --output ./src/partials.cjs -c handlebars/runtime && handlebars ./templates -e hbs --output ./src/templates.cjs -c handlebars/runtime",
+    "prepare": "pnpm run build",
+    "prepublish": "pnpm run build"
   },
+  "type": "module",
   "keywords": [
     "jsonresume",
     "jsonresume-theme",
@@ -27,12 +25,17 @@
   "license": "MIT",
   "dependencies": {
     "handlebars": "^4",
+    "lodash": "^4.17.21",
     "moment": "^2.19.1",
     "pluralize": "^7.0.0"
   },
   "devDependencies": {
+    "@babel/core": "^7.23.5",
+    "@babel/preset-env": "^7.23.5",
+    "babel-loader": "^9.1.3",
     "css-loader": "^6.8.1",
     "handlebars-loader": "^1.7.3",
+    "parcel": "^2.10.3",
     "raw-loader": "^4.0.2",
     "style-loader": "^3.3.3",
     "webpack": "^5.89.0",
