+++ themes/papirus/partials/summary.hbs
