--- themes/papirus/src/formatDateYear.js
@@ -1,7 +1,0 @@
-import moment from 'moment';
-export default function (date) {
-  if (typeof date === 'undefined') {
-    return 'now';
-  }
-  return moment(date).format('YYYY');
-}
