+++ themes/papirus/src/index.js
@@ -1,8 +1,15 @@
-import css from '!!raw-loader!./style.css';
-import tpl from './resume.hbs';
+import * as helpers from './helpers.js';
+import Handlebars from 'handlebars/runtime.js';
+import './partials.cjs';
+import './templates.cjs';
+import css from './style.js';
 
-export const render = (resume) =>
-  tpl({
-    css: css,
-    resume: resume,
-  });
+export const render = async (resume) => {
+  return Handlebars.templates['resume'](
+    {
+      css: css,
+      resume: resume,
+    },
+    { helpers }
+  );
+};
