--- themes/papirus/src/networkIcon.js
@@ -1,7 +1,0 @@
-export default function (network) {
-  if (network === 'StackOverflow') {
-    return 'stack-overflow';
-  } else {
-    return network.toLowerCase();
-  }
-}
