--- themes/papirus/src/wordWrap.js
@@ -1,4 +1,0 @@
-export default function (str) {
-  str = str.replace(/\//g, '/ ');
-  return str.replace('/ / ', '//');
-}
